pyproject.toml
[project]
name = "resumeup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google>=3.0.0",
    "google-generativeai>=0.8.5",
    "openai>=1.84.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
